requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
altair==5.1.2
annotated-types==0.6.0
anthropic==0.25.6
anyio==3.7.1
async-timeout==4.0.3
attrs==23.1.0
azure-ai-formrecognizer==3.3.1
azure-common==1.1.28
azure-core==1.29.4
backoff==2.2.1
beautifulsoup4==4.12.3
black==23.10.0
blinker==1.6.3
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer==3.3.0
click==8.1.7
cohere==4.44
colorlog==6.8.0
contourpy==1.1.1
cycler==0.12.1
dataclasses-json==0.6.3
distro==1.9.0
dnspython==2.4.2
entrypoints==0.4
exceptiongroup==1.1.3
Faker==19.10.0
fastapi==0.104.1
fastavro==1.9.3
favicon==0.7.0
filelock==3.13.1
fonttools==4.43.1
frozenlist==1.4.0
fsspec==2023.12.2
gitdb==4.0.10
GitPython==3.1.38
googleapis-common-protos==1.62.0
greenlet==3.0.1
grpc-gateway-protoc-gen-openapiv2==0.1.0
grpcio==1.60.0
h11==0.14.0
htbuilder==0.6.2
httpcore==1.0.2
httpx==0.26.0
idna==3.4
importlib-metadata==6.8.0
isodate==0.6.1
Jinja2==3.1.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.19.1
jsonschema-specifications==2023.7.1
kiwisolver==1.4.5
langchain==0.0.305
langchain-core==0.1.14
langchain-openai==0.0.3
langsmith==0.0.83
loguru==0.7.2
lxml==4.9.3
lz4==4.3.2
Markdown==3.5
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
more-itertools==10.1.0
mpmath==1.3.0
msrest==0.7.1
multidict==6.0.4
mypy-extensions==1.0.0
networkx==3.2.1
numexpr==2.8.7
numpy==1.26.1
oauthlib==3.2.2
openai==1.9.0
packaging==23.2
pandas==2.1.1
pathspec==0.11.2
Pillow==10.1.0
pinecone-client==2.2.4
platformdirs==3.11.0
protobuf==3.20.3
pyarrow==13.0.0
pydantic==1.10.14
pydantic_core==2.10.1
pydeck==0.8.1b0
Pygments==2.16.1
pymdown-extensions==10.3
pyparsing==3.1.1
pypdf==4.1.0
PyPDF2==3.0.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.30.2
regex==2023.10.3
requests==2.31.0
requests-oauthlib==1.3.1
rich==13.6.0
rpds-py==0.10.6
semantic-router==0.0.17
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.23
st-annotated-text==4.0.1
starlette==0.27.0
streamlit==1.27.2
streamlit-camera-input-live==0.2.0
streamlit-card==0.0.61
streamlit-embedcode==0.1.2
streamlit-extras==0.3.4
streamlit-faker==0.0.2
streamlit-image-coordinates==0.1.6
streamlit-javascript==0.1.5
streamlit-keyup==0.2.0
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==1.0.2
sympy==1.12
tenacity==8.2.3
tiktoken==0.5.2
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.1.2
tornado==6.3.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.1
urllib3==2.0.6
uvicorn==0.24.0.post1
validators==0.22.0
watchdog==3.0.0
yarl==1.9.2
zipp==3.17.0
